+++ FrontEnd/coded/src/component/Footer/Footer.js
