+++ FrontEnd/coded/src/component/Footer/Footer.scss
